Config is only saved if you don't have a .latemprc file already

+++ install.sh
@@ -9,6 +9,10 @@ echo "Installing latemp in your /usr/local/bin/ folder..."
 chmod +x ./latemp;
 echo "I highly recommend reading through the script before installing it"
 sudo cp ./latemp /usr/local/bin/
-cp ./config $HOME/.latemprc
+if [ -f "$HOME/.latemprc" ]; then
+	:
+else
+	cp ./config $HOME/.latemprc
+fi
 echo "---------------------------------------------------"
 echo "-You can now use it by typing latemp in your shell-"
